services/orchestrator/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-telegram-bot==20.7
aiohttp==3.9.1
pydantic==2.5.0
psutil==5.9.6
asyncpg==0.29.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
sentence-transformers==2.7.0
numpy==1.24.3
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
websockets==12.0
beautifulsoup4==4.12.3
duckduckgo-search==5.3.0
google-generativeai==0.8.3
